mesaj

+++ follow-the-beat/src/main/java/com/maco/followthebeat/v2/concertmatcher/MatchResult.java
@@ -3,5 +3,5 @@ package com.maco.followthebeat.v2.concertmatcher;
 public class MatchResult {
     public String concertId;
     public String name;
-    public double score;
+    public float score;
 }
